requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
DateTime==4.4
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
idna==3.3
kombu==5.2.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
razorpay==1.3.0
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
